requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
Pillow>=10.0.0
easyocr>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
chromadb>=0.4.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
langgraph>=0.0.40
pydantic>=2.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
httpx>=0.24.0
openai>=1.0.0
torchvision>=0.15.0
